+++ tests/testthat/test-snapshot-file.R
@@ -9,14 +9,13 @@ test_that("expect_snapshot_file works", {
   expect_snapshot_file(path, "foo.png")
 
   path <- withr::local_tempfile()
-  mtcars2 <- mtcars
-  # mtcars2$wt[10] <- NA
+  mtcars2 <- mtcars[1:5, 1:4]
   write.csv(mtcars2, path)
   expect_snapshot_file(path, "foo.csv", compare = compare_file_text)
 
   # Deprecated `binary` argument still works
   withr::local_options(lifecycle_verbosity = "quiet")
-  expect_snapshot_file(path, "foo.csv", binary = FALSE)
+  expect_snapshot_file(path, "foo-not-binary.csv", binary = FALSE)
 })
 
 
@@ -39,8 +38,19 @@ test_that("expect_snapshot_file works with variant", {
   )
 })
 
+test_that("expect_snapshot_file finds duplicate snapshot files", {
+  expect_snapshot(
+    expect_snapshot_file(
+      write_tmp_lines(r_version()),
+      "version.txt",
+      variant = r_version()
+    ),
+    error = TRUE
+  )
+})
+
 test_that("basic workflow", {
-  snapper <- local_snapshotter(fail_on_new = FALSE)
+  snapper <- local_test_snapshotter()
 
   path <- write_tmp_lines(letters)
   # warns on first run
@@ -61,7 +71,7 @@ test_that("basic workflow", {
 })
 
 test_that("can announce snapshot file", {
-  snapper <- local_snapshotter(fail_on_new = FALSE)
+  snapper <- local_test_snapshotter()
   snapper$start_file("snapshot-announce", "test")
   announce_snapshot_file(name = "bar.svg")
   expect_equal(snapper$snap_file_seen, "snapshot-announce/bar.svg")
@@ -151,6 +161,7 @@ test_that("snapshot_hint output differs in R CMD check", {
   snapshot_review_hint <- function(...) {
     testthat:::snapshot_review_hint(..., reset_output = FALSE)
   }
+  withr::local_envvar(GITHUB_ACTIONS = "false")
 
   expect_snapshot(cat(snapshot_review_hint(
     "lala",
@@ -164,6 +175,18 @@ test_that("snapshot_hint output differs in R CMD check", {
     check = TRUE,
     ci = FALSE
   )))
+  expect_snapshot(cat(snapshot_review_hint(
+    "lala",
+    "foo.r",
+    check = TRUE,
+    ci = TRUE
+  )))
+
+  withr::local_envvar(
+    GITHUB_ACTIONS = "true",
+    GITHUB_REPOSITORY = "r-lib/testthat",
+    GITHUB_RUN_ID = "123"
+  )
   expect_snapshot(cat(snapshot_review_hint(
     "lala",
     "foo.r",